backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow

huggingface-hub
requests==2.31.0
pdf2image==1.16.3
python-dotenv==1.0.0

# LangChain and LangFuse integration
langchain==0.3.7
langchain-huggingface==0.1.2
langchain-core==0.3.15
langfuse<3.0.0
